aadikarta-social-media/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlmodel
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-multipart
structlog
minio
openai
requests
python-dotenv
tenacity
pathlib
requests-oauthlib
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
facebook-sdk
apscheduler
slowapi
prometheus-fastapi-instrumentator
aiohttp
replicate
